+++ service-products/src/main/java/servicesproductos/entities/ProductoDTO.java
@@ -14,6 +14,6 @@ import lombok.*;
 public class ProductoDTO {
     private Long id;
     private String nombre;
-    private Double valor;
+    private int valor;
     private String categoriaNombre;
 }
